Update the NMTK setup so that we add in a local_settings.py flag that allows us to enable/disable the tool server (or NMTK server) built into this setup

+++ NMTK_apps/NMTK_server/management/commands/purge_users.py
@@ -1,8 +1,11 @@
 from django.core.management.base import BaseCommand, CommandError
 from django.contrib.auth.models import User
+from django.conf import settings
 
 class Command(BaseCommand):
     help = 'Purge an existing user from the database.'
-    def handle(self, *args, **options):        
+    def handle(self, *args, **options):  
+        if not settings.NMTK_SERVER:
+            raise CommandError('The NMTK Server is not currently enabled')      
         for user in args:
             User.objects.filter(username=user).delete()
